refactor(SafeRedirect): remove whitelist into utils
